+++ HospitalBudget-ejb/ejbModule/cn/dmdl/stl/hospitalbudget/budget/entity/ExpendConfirmInfo.java
@@ -28,7 +28,7 @@ public class ExpendConfirmInfo implements java.io.Serializable{
 	private Date insertTime;
 	private Date updateTime;
 	private Integer confirmUser;
-	private Float totalMoney;//总金额
+	private Double totalMoney;//总金额
 	private Date confirm_time;//确认时间
 	private Integer confirm_status;//确认单状态 0初始状态 1确认完成 2 确认驳回
 	private boolean deleted;//删除
@@ -90,10 +90,10 @@ public class ExpendConfirmInfo implements java.io.Serializable{
 	
 	@Column(name = "total_money")
 	@NotNull
-	public Float getTotalMoney() {
+	public Double getTotalMoney() {
 		return totalMoney;
 	}
-	public void setTotalMoney(Float totalMoney) {
+	public void setTotalMoney(Double totalMoney) {
 		this.totalMoney = totalMoney;
 	}
